Moved UI code out of dataset reading code.
