+++ FMEAR/ViewController+Document.swift
@@ -13,116 +13,116 @@ import SceneKit.ModelIO
 
 extension ViewController: FileManagerDelegate {
 
-    func openDataset(url: URL) {
-        print("openDataset")
-        
-        if datasets[url] == nil {
-            // We want to extract the data from the document url and create a scene node.
-            let document = Document(fileURL: url)
-            document.open() { success in
-                if success {
-                    // Record the dataset
-                    let dataset = Dataset()
-                    dataset.documentURL = url
-                    self.datasets[url] = dataset
-                    
-                    let documentName = document.fileURL.pathComponents.last ?? ""
-                    //self.textManager.showMessage("Opening \(documentName)...")
-
-                    // Create a file manager to handle create and remove folders
-                    let fileManager = FileManager.default
-                    fileManager.delegate = self
-                    
-                    // Get the document directory path
-                    guard let documentDirectory = fileManager.urls(for: .documentDirectory, in: .userDomainMask).first as URL? else {
-                        self.errors.append(FMEError.failedToOpenDataset(url))
-                        return
-                    }
-                    
-                    // Create the destination folder for the unzipped content
-                    let unzippedFolderUrl = documentDirectory.appendingPathComponent(UUID().uuidString)
-                    do {
-                        try fileManager.createDirectory(at: unzippedFolderUrl,
-                                                        withIntermediateDirectories: true,
-                                                        attributes: nil)
-                    } catch {
-                        self.errors.append(FMEError.failedToOpenDataset(url))
-                        return
-                    }
-                    
-                    // Unzip the .fmear file
-                    if let unzipError = self.unzipFile(fileUrl: url, destinationFolderUrl: unzippedFolderUrl) {
-                        self.errors.append(unzipError)
-                    } else {
-                        // Load the model into the scene
-                        dataset.settings = self.readSettings(folderUrl: unzippedFolderUrl)
-                        dataset.model = self.readModel(folderUrl: unzippedFolderUrl,
-                                                       settings: dataset.settings)
-                                                
-//                        if let model = dataset.model {
-//                            self.logSceneNode(model, level: 0)
+//    func openDataset2(url: URL) {
+//        print("openDataset2")
+//
+//        if datasets[url] == nil {
+//            // We want to extract the data from the document url and create a scene node.
+//            let document = Document(fileURL: url)
+//            document.open() { success in
+//                if success {
+//                    // Record the dataset
+//                    let dataset = Dataset()
+//                    dataset.documentURL = url
+//                    self.datasets[url] = dataset
+//
+//                    let documentName = document.fileURL.pathComponents.last ?? ""
+//                    //self.textManager.showMessage("Opening \(documentName)...")
+//
+//                    // Create a file manager to handle create and remove folders
+//                    let fileManager = FileManager.default
+//                    fileManager.delegate = self
+//
+//                    // Get the document directory path
+//                    guard let documentDirectory = fileManager.urls(for: .documentDirectory, in: .userDomainMask).first as URL? else {
+//                        self.errors.append(FMEError.failedToOpenDataset(url))
+//                        return
+//                    }
+//
+//                    // Create the destination folder for the unzipped content
+//                    let unzippedFolderUrl = documentDirectory.appendingPathComponent(UUID().uuidString)
+//                    do {
+//                        try fileManager.createDirectory(at: unzippedFolderUrl,
+//                                                        withIntermediateDirectories: true,
+//                                                        attributes: nil)
+//                    } catch {
+//                        self.errors.append(FMEError.failedToOpenDataset(url))
+//                        return
+//                    }
+//
+//                    // Unzip the .fmear file
+//                    if let unzipError = self.unzipFile(fileUrl: url, destinationFolderUrl: unzippedFolderUrl) {
+//                        self.errors.append(unzipError)
+//                    } else {
+//                        // Load the model into the scene
+//                        dataset.settings = self.readSettings(folderUrl: unzippedFolderUrl)
+//                        dataset.model = self.readModel(folderUrl: unzippedFolderUrl,
+//                                                       settings: dataset.settings)
+//
+////                        if let model = dataset.model {
+////                            self.logSceneNode(model, level: 0)
+////                        }
+//
+//                        // If there is a model, we will add it to the scene in the next
+//                        // frame update.
+//                        if dataset.model != nil {
+//                            self.datasetsReady.append(url)
 //                        }
-                        
-                        // If there is a model, we will add it to the scene in the next
-                        // frame update.
-                        if dataset.model != nil {
-                            self.datasetsReady.append(url)
-                        }
-                    }
-                    
-                    // Remove the unzipped folder
-                    do {
-                        try fileManager.removeItem(at: unzippedFolderUrl)
-                    } catch let error {
-                        print("Failed to remove the unzipped content in \(unzippedFolderUrl.path): \(error)")
-                    }
-
-                    // Close the document since we don't need it anymore
-                    document.close() { success in
-                        if success {
-                            print("\(documentName) closed")
-                        } else {
-                            print("Failed to close \(documentName)")
-                        }
-                    }
-                    
-                    // Update the compass.
-                    // TODO: Currently we replace the compass image. If in the future
-                    // we want to handle more than one models, we will need to update
-                    // the image that includes all models.
-                    
-                    
-                } else {
-                    self.errors.append(FMEError.failedToOpenDataset(url))
-                }
-            }
-
-        } else {
-            // TODO: We have opened this dataset. We will reuse the model.
-
-            // We do nothing now
-        }
-    }
-    
-    func unzipFile(fileUrl: URL, destinationFolderUrl: URL) -> Error? {
-        if SSZipArchive.unzipFile(atPath: fileUrl.path, toDestination: destinationFolderUrl.path) {
-            return nil
-        } else {
-            return FMEError.failedToOpenDataset(fileUrl)
-        }
-    }
-    
-    func closeDataset(url: URL) {
-        datasets.removeValue(forKey: url)
-    }
-    
-    func reloadAllDatasets() {
-        let keys = datasets.keys
-        datasets.removeAll()
-        for key in keys {
-            openDataset(url: key)
-        }
-    }
+//                    }
+//
+//                    // Remove the unzipped folder
+//                    do {
+//                        try fileManager.removeItem(at: unzippedFolderUrl)
+//                    } catch let error {
+//                        print("Failed to remove the unzipped content in \(unzippedFolderUrl.path): \(error)")
+//                    }
+//
+//                    // Close the document since we don't need it anymore
+//                    document.close() { success in
+//                        if success {
+//                            print("\(documentName) closed")
+//                        } else {
+//                            print("Failed to close \(documentName)")
+//                        }
+//                    }
+//
+//                    // Update the compass.
+//                    // TODO: Currently we replace the compass image. If in the future
+//                    // we want to handle more than one models, we will need to update
+//                    // the image that includes all models.
+//
+//
+//                } else {
+//                    self.errors.append(FMEError.failedToOpenDataset(url))
+//                }
+//            }
+//
+//        } else {
+//            // TODO: We have opened this dataset. We will reuse the model.
+//
+//            // We do nothing now
+//        }
+//    }
+//
+//    func unzipFile(fileUrl: URL, destinationFolderUrl: URL) -> Error? {
+//        if SSZipArchive.unzipFile(atPath: fileUrl.path, toDestination: destinationFolderUrl.path) {
+//            return nil
+//        } else {
+//            return FMEError.failedToOpenDataset(fileUrl)
+//        }
+//    }
+//
+//    func closeDataset(url: URL) {
+//        datasets.removeValue(forKey: url)
+//    }
+//
+//    func reloadAllDatasets() {
+//        let keys = datasets.keys
+//        datasets.removeAll()
+//        for key in keys {
+//            openDataset(url: key)
+//        }
+//    }
     
 //    func openDocument(document: UIDocument) {
 //        // Access the document
@@ -200,211 +200,191 @@ extension ViewController: FileManagerDelegate {
 //        })
 //    }
     
-    func readSettings(folderUrl: URL) -> Settings? {
-        let fileManager = FileManager.default
-        fileManager.delegate = self
-        guard let enumerator = fileManager.enumerator(atPath: folderUrl.path) else {
-            return nil
-        }
-        
-        // Find the settings.json file
-        while let objectPath = enumerator.nextObject() as? String {
-            if objectPath.hasPrefix("__MACOSX") {
-                // Ignore this __MACOSX folder
-                continue
-            }
-            
-            if objectPath.hasSuffix("settings.json") {
-                return readSettingsFile(file: folderUrl.appendingPathComponent(objectPath))
-            } else if objectPath.hasSuffix(".json") {
-                // Version 1 and 2 of the settings json file has a "model" name
-                // that should match the folder name inside the .fmear archive
-                let jsonPath = folderUrl.appendingPathComponent(objectPath)
-                let jsonFilename = jsonPath.deletingPathExtension().lastPathComponent
-                let folderName = jsonPath.deletingLastPathComponent().lastPathComponent
-                if jsonFilename == folderName {
-                    return readSettingsFile(file: jsonPath)
-                }
-            }
-        }
-        
-        return nil
-    }
+//    func readSettings(folderUrl: URL) -> Settings? {
+//        let fileManager = FileManager.default
+//        fileManager.delegate = self
+//        guard let enumerator = fileManager.enumerator(atPath: folderUrl.path) else {
+//            return nil
+//        }
+//
+//        // Find the settings.json file
+//        while let objectPath = enumerator.nextObject() as? String {
+//            if objectPath.hasPrefix("__MACOSX") {
+//                // Ignore this __MACOSX folder
+//                continue
+//            }
+//
+//            if objectPath.hasSuffix("settings.json") {
+//                return readSettingsFile(file: folderUrl.appendingPathComponent(objectPath))
+//            } else if objectPath.hasSuffix(".json") {
+//                // Version 1 and 2 of the settings json file has a "model" name
+//                // that should match the folder name inside the .fmear archive
+//                let jsonPath = folderUrl.appendingPathComponent(objectPath)
+//                let jsonFilename = jsonPath.deletingPathExtension().lastPathComponent
+//                let folderName = jsonPath.deletingLastPathComponent().lastPathComponent
+//                if jsonFilename == folderName {
+//                    return readSettingsFile(file: jsonPath)
+//                }
+//            }
+//        }
+//
+//        return nil
+//    }
     
-    func readModel(folderUrl: URL, settings: Settings?) -> SCNNode? {
-        let fileManager = FileManager.default
-        fileManager.delegate = self
-        guard let enumerator = fileManager.enumerator(atPath: folderUrl.path) else {
-            return nil
-        }
-        
-        // Find the settings.json file
-        var assets: [SCNNode] = []
-        while let objectPath = enumerator.nextObject() as? String {
-            if objectPath.hasPrefix("__MACOSX") {
-                // Ignore this __MACOSX folder
-                continue
-            }
-            
-            let objSuffix = ".obj"
-            if objectPath.hasSuffix(objSuffix) {
-                if let objNode = readObjFile(file: folderUrl.appendingPathComponent(objectPath)) {
-                    // Set the node name as the OBJ file name, which should
-                    // be the asset/feature type name from the FME AR writer
-                    var assetName = objectPath
-                    assetName.removeLast(objSuffix.count)
-                    print("Reading asset \(assetName)")
-                    objNode.name = assetName
-                    assets.append(objNode)
-                }
-            }
-        }
-        
-        print("\(assets.count) asset(s) found")
-        
-        if (assets.count > 0) {
-            
-            let model = SCNNode()
-            for asset in assets {
-                model.addChildNode(asset)
-            }
-            
-            self.overlayView.compass().image
-                = model.snapshot(size: CGSize(width: 512.0, height: 512.0))
-
-            // Dimension and position
-            let (minCoord, maxCoord) = model.boundingBox
-            let centerX = (minCoord.x + maxCoord.x) * 0.5
-            let centerY = (minCoord.y + maxCoord.y) * 0.5
-            let groundZ = 0.0
-            
-            // json.settings version 4 - Viewpoints
-            // ----------
-            let viewpoints = settings?.viewpoints ?? []
-                        
-            // json.settings version 3 - Anchor
-            // ------
-            // By default, set the anchor to the center of the model, with the
-            // 0.0 height as the ground
-            var anchor: SCNVector3 = SCNVector3(centerX, Float(groundZ), centerY) // default
-            var isDefaultAnchor = true
-            if let anchors = settings?.anchors {
-                if let firstAnchor = anchors.first {
-                    
-                    if firstAnchor.x != nil && firstAnchor.y != nil {
-                        isDefaultAnchor = false
-                    }
-                    
-                    anchor = SCNVector3(firstAnchor.x ?? Double(centerX),
-                                       firstAnchor.z ?? groundZ,
-                                       firstAnchor.y ?? Double(centerY))
-                }
-            }
-            
-            // json.settings version 3
-            if viewpoints.isEmpty {
-                self.serialQueue.async {
-                    let viewpointLabelNode = self.overlayView.labelNode(labelName: self.viewpointLabelName,
-                                                                        iconNamed: LabelIcons.geolocationAnchor.rawValue)
-                    viewpointLabelNode.isHidden = !(UserDefaults.standard.bool(for: .drawAnchor))
-                    viewpointLabelNode.callToAction = false
-
-                    if isDefaultAnchor {
-                        viewpointLabelNode.text = "Anchor (Default)"
-                    } else {
-                        viewpointLabelNode.text = "Anchor (Custom)"
-                    }
-                }
-                
-                // Position the container node, including the model and the anchor
-                // node, to the anchor location.
-                // The FME coordinate z axis = ARKit y axis
-                // The FME coordinate y axis = ARKit z axis
-                model.position = SCNVector3(-anchor.x, -anchor.y, anchor.z)
-            } else {
-                // If we have a viewpont, we should always show it at the beginning
-                UserDefaults.standard.set(true, for: .drawAnchor)
-            }
-            
-            // Rotate to Y up
-            model.eulerAngles.x = -Float.pi / 2
-            
-            // TODO: Change the name to something more meaningful
-            model.name = "VirtualObjectContent"
-            
-            let modelDimension = self.dimension(model)
-            let maxLength = max(modelDimension.x, modelDimension.y, modelDimension.z)
-            let definition = VirtualObjectDefinition(modelName: "model", displayName: "model", particleScaleInfo: [:])
-            let object = VirtualObject(definition: definition,
-                                     modelNode: model,
-                                     viewpoints: viewpoints)
-            
-            // TODO: Change the name to something more meaningful
-            object.name = "VirtualObject"
-
-            if let firstViewpoint = object.viewpoints.first {
-                object.anchorAtViewpoint(viewpointId: firstViewpoint.id)
-            }
-            
-            // Set a cteagory bit mask to include the virtual object in the hit test.
-            object.categoryBitMask = HitTestOptionCategoryBitMasks.virtualObject.rawValue
-            
-            // Scale the virtual object
-            if maxLength > 0 {
-                // By default, scale the model to be within a 0.5 meter cube.
-                // If the scaling is set in the model json file, use it instead.
-                if self.scaleMode == .fullScale {
-                    object.scale = SCNVector3(1.0, 1.0, 1.0)
-                } else if let userSpecifiedScale = self.scaling {
-                    object.scale = SCNVector3(userSpecifiedScale, userSpecifiedScale, userSpecifiedScale)
-                } else {
-                    let preferredScale = (Float(0.5) / maxLength)
-                    object.scale = SCNVector3(preferredScale, preferredScale, preferredScale)
-                }
-                
-                // Set scale lock
-                self.virtualObjectManager.allowScaling = !self.scaleLockEnabled
-            }
-            
-            return object
-        } else {
-            return nil
-        }
-    }
-    
-    func readObjFile(file: URL) -> SCNNode? {
-        let loadingOptions = [
-            SCNSceneSource.LoadingOption.createNormalsIfAbsent : false,
-            SCNSceneSource.LoadingOption.convertToYUp: false,
-            SCNSceneSource.LoadingOption.flattenScene: true]
-        
-        if let sceneSource = SCNSceneSource(url: file, options: loadingOptions) {
-            //self.logSceneSource(sceneSource)
-            
-            let statusHandler = { (totalProgress: Float, status: SCNSceneSourceStatus, error: Error?, stopLoading: UnsafeMutablePointer<ObjCBool>) -> Void in
-                switch status {
-                case .error: print("error: \(totalProgress)")
-                case .parsing: print("parsing: \(totalProgress)")
-                case .validating: print("validating: \(totalProgress)")
-                case .processing: print("processing: \(totalProgress)")
-                case .complete: print("complete: \(totalProgress)")
-                default: print("default status: \(totalProgress)");
-                }
-            };
-            
-            if let scene = sceneSource.scene(options: loadingOptions, statusHandler:  statusHandler) {
-                
-                // TODO: SceneKit gives an error (Removing the root node
-                // of a scene from its scene is not allowed), but cloning
-                // doesn't work. The clone seems to lose the material colours
-                let rootNode = scene.rootNode
-                adjustMaterialProperties(sceneNode: rootNode)
-                return rootNode
-            }
-        }
-        
-        return nil
-    }
+//    func readModel(folderUrl: URL, settings: Settings?) -> SCNNode? {
+//        let fileManager = FileManager.default
+//        fileManager.delegate = self
+//        guard let enumerator = fileManager.enumerator(atPath: folderUrl.path) else {
+//            return nil
+//        }
+//
+//        // Find the settings.json file
+//        var assets: [SCNNode] = []
+//        while let objectPath = enumerator.nextObject() as? String {
+//            if objectPath.hasPrefix("__MACOSX") {
+//                // Ignore this __MACOSX folder
+//                continue
+//            }
+//
+//            let objSuffix = ".obj"
+//            if objectPath.hasSuffix(objSuffix) {
+//                if let objNode = readObjFile(file: folderUrl.appendingPathComponent(objectPath)) {
+//                    // Set the node name as the OBJ file name, which should
+//                    // be the asset/feature type name from the FME AR writer
+//                    var assetName = objectPath
+//                    assetName.removeLast(objSuffix.count)
+//                    print("Reading asset \(assetName)")
+//                    objNode.name = assetName
+//                    assets.append(objNode)
+//                }
+//            }
+//        }
+//
+//        print("\(assets.count) asset(s) found")
+//
+//        if (assets.count > 0) {
+//
+//            let model = SCNNode()
+//            for asset in assets {
+//                model.addChildNode(asset)
+//            }
+//
+//            // Dimension and position
+//            let (minCoord, maxCoord) = model.boundingBox
+//            let centerX = (minCoord.x + maxCoord.x) * 0.5
+//            let centerY = (minCoord.y + maxCoord.y) * 0.5
+//            let groundZ = 0.0
+//
+//            // json.settings version 4 - Viewpoints
+//            // ----------
+//            let viewpoints = settings?.viewpoints ?? []
+//
+//            // json.settings version 3 - Anchor
+//            // ------
+//            // By default, set the anchor to the center of the model, with the
+//            // 0.0 height as the ground
+//            var anchor: SCNVector3 = SCNVector3(centerX, Float(groundZ), centerY) // default
+//            if let anchors = settings?.anchors {
+//                if let firstAnchor = anchors.first {
+//                    anchor = SCNVector3(firstAnchor.x ?? Double(centerX),
+//                                       firstAnchor.z ?? groundZ,
+//                                       firstAnchor.y ?? Double(centerY))
+//                }
+//            }
+//
+//            // json.settings version 3
+//            if viewpoints.isEmpty {
+//                // Position the container node, including the model and the anchor
+//                // node, to the anchor location.
+//                // The FME coordinate z axis = ARKit y axis
+//                // The FME coordinate y axis = ARKit z axis
+//                model.position = SCNVector3(-anchor.x, -anchor.y, anchor.z)
+//            } else {
+//                // If we have a viewpont, we should always show it at the beginning
+//                UserDefaults.standard.set(true, for: .drawAnchor)
+//            }
+//
+//            // Rotate to Y up
+//            model.eulerAngles.x = -Float.pi / 2
+//
+//            // TODO: Change the name to something more meaningful
+//            model.name = "VirtualObjectContent"
+//
+//            let modelDimension = self.dimension(model)
+//            let maxLength = max(modelDimension.x, modelDimension.y, modelDimension.z)
+//            let definition = VirtualObjectDefinition(modelName: "model", displayName: "model", particleScaleInfo: [:])
+//            let object = VirtualObject(definition: definition,
+//                                     modelNode: model,
+//                                     viewpoints: viewpoints)
+//
+//            // TODO: Change the name to something more meaningful
+//            object.name = "VirtualObject"
+//
+//            if let firstViewpoint = object.viewpoints.first {
+//                object.anchorAtViewpoint(viewpointId: firstViewpoint.id)
+//            }
+//
+//            // Set a cteagory bit mask to include the virtual object in the hit test.
+//            object.categoryBitMask = HitTestOptionCategoryBitMasks.virtualObject.rawValue
+//
+//            // Scale the virtual object
+//            if maxLength > 0 {
+//                // By default, scale the model to be within a 0.5 meter cube.
+//                // If the scaling is set in the model json file, use it instead.
+//                if self.scaleMode == .fullScale {
+//                    object.scale = SCNVector3(1.0, 1.0, 1.0)
+//                } else if let userSpecifiedScale = self.scaling {
+//                    object.scale = SCNVector3(userSpecifiedScale, userSpecifiedScale, userSpecifiedScale)
+//                } else {
+//                    let preferredScale = (Float(0.5) / maxLength)
+//                    object.scale = SCNVector3(preferredScale, preferredScale, preferredScale)
+//                }
+//
+//                // Set scale lock
+//                DispatchQueue.main.async {
+//                    self.virtualObjectManager.allowScaling = !self.scaleLockEnabled
+//                }
+//            }
+//
+//            return object
+//        } else {
+//            return nil
+//        }
+//    }
+//
+//    func readObjFile(file: URL) -> SCNNode? {
+//        let loadingOptions = [
+//            SCNSceneSource.LoadingOption.createNormalsIfAbsent : false,
+//            SCNSceneSource.LoadingOption.convertToYUp: false,
+//            SCNSceneSource.LoadingOption.flattenScene: true]
+//
+//        if let sceneSource = SCNSceneSource(url: file, options: loadingOptions) {
+//            //self.logSceneSource(sceneSource)
+//
+//            let statusHandler = { (totalProgress: Float, status: SCNSceneSourceStatus, error: Error?, stopLoading: UnsafeMutablePointer<ObjCBool>) -> Void in
+//                switch status {
+//                case .error: print("error: \(totalProgress)")
+//                case .parsing: print("parsing: \(totalProgress)")
+//                case .validating: print("validating: \(totalProgress)")
+//                case .processing: print("processing: \(totalProgress)")
+//                case .complete: print("complete: \(totalProgress)")
+//                default: print("default status: \(totalProgress)");
+//                }
+//            };
+//
+//            if let scene = sceneSource.scene(options: loadingOptions, statusHandler:  statusHandler) {
+//
+//                // TODO: SceneKit gives an error (Removing the root node
+//                // of a scene from its scene is not allowed), but cloning
+//                // doesn't work. The clone seems to lose the material colours
+//                let rootNode = scene.rootNode
+//                adjustMaterialProperties(sceneNode: rootNode)
+//                return rootNode
+//            }
+//        }
+//
+//        return nil
+//    }
     
 //    func closeDocument(document: UIDocument) {
 //        print("Closing '\(document.fileURL)'...")
@@ -433,40 +413,40 @@ extension ViewController: FileManagerDelegate {
 ////        documentOpened = false
 //    }
     
-    func readSettingsFile(file: URL) -> Settings? {
-        
-        var settings: Settings?
-        
-        do {
-            let jsonData = try Data(contentsOf: file)
-            let jsonDict = try JSONSerialization.jsonObject(with: jsonData, options: [])
-            settings = try Settings(json: jsonDict)
-        } catch {
-            print("No settings")
-            settings = nil
-        }
-        
-        // TODO: Move the following state-changing code to somewhere else
-        self.scaleMode = .customScale
-        self.scaleLockEnabled = false
-        if let scaling = settings?.scaling {
-            if scaling == 1.0 {
-                self.scaleMode = .fullScale
-            } else {
-                self.scaleMode = .customScale
-            }
-
-            self.scaleLockEnabled = true
-            self.scaling = scaling
-        }
-
-        DispatchQueue.main.async {                   
-            // Update the scale options button
-            self.setShowScaleOptionsButton(mode: self.scaleMode, lockOn: self.scaleLockEnabled)
-        }
-        
-        return settings
-    }
+//    func readSettingsFile(file: URL) -> Settings? {
+//
+//        var settings: Settings?
+//
+//        do {
+//            let jsonData = try Data(contentsOf: file)
+//            let jsonDict = try JSONSerialization.jsonObject(with: jsonData, options: [])
+//            settings = try Settings(json: jsonDict)
+//        } catch {
+//            print("No settings")
+//            settings = nil
+//        }
+//
+//        // TODO: Move the following state-changing code to somewhere else
+//        self.scaleMode = .customScale
+//        self.scaleLockEnabled = false
+//        if let scaling = settings?.scaling {
+//            if scaling == 1.0 {
+//                self.scaleMode = .fullScale
+//            } else {
+//                self.scaleMode = .customScale
+//            }
+//
+//            self.scaleLockEnabled = true
+//            self.scaling = scaling
+//        }
+//
+//        DispatchQueue.main.async {
+//            // Update the scale options button
+//            self.setShowScaleOptionsButton(mode: self.scaleMode, lockOn: self.scaleLockEnabled)
+//        }
+//
+//        return settings
+//    }
     
 //    func loadModel(path: URL) {
 //
@@ -751,181 +731,181 @@ extension ViewController: FileManagerDelegate {
 //        //self.modelPath = nil;
 //    }
     
-    func adjustMaterialProperties(sceneNode: SCNNode) {
-        
-        if let geometry = sceneNode.geometry {
-            for material in geometry.materials {
-
-                // FMEMOBILE-384
-                // SCNSceneSource seems to mistakenly load the OBJ Ka (ambient)
-                // material property as emission property. Since we dont' care
-                // emission for now, we will just copy the emission value to the
-                // original ambient value, and reset the emission property
-                // to zero.
-                material.ambient.contents = material.emission.contents
-                material.emission.contents = UIColor(red: 0.0, green: 0.0, blue: 0.0, alpha: 1.0)
-                
-                // Set doubleSided so that the user can always see the model
-                material.isDoubleSided = true
-                
-                // For some reasons, the Tr value in the OBJ model is not read
-                // correctly using SCNSceneSource or Model IO. If the OBJ material has
-                // the d value.
-                if let transparent = material.transparent.contents as? NSNumber {
-                    material.transparency = CGFloat(transparent.floatValue)
-                }
-            }
-
-        }
-        
-        for childNode in sceneNode.childNodes {
-            adjustMaterialProperties(sceneNode: childNode)
-        }
-    }
+//    func adjustMaterialProperties(sceneNode: SCNNode) {
+//
+//        if let geometry = sceneNode.geometry {
+//            for material in geometry.materials {
+//
+//                // FMEMOBILE-384
+//                // SCNSceneSource seems to mistakenly load the OBJ Ka (ambient)
+//                // material property as emission property. Since we dont' care
+//                // emission for now, we will just copy the emission value to the
+//                // original ambient value, and reset the emission property
+//                // to zero.
+//                material.ambient.contents = material.emission.contents
+//                material.emission.contents = UIColor(red: 0.0, green: 0.0, blue: 0.0, alpha: 1.0)
+//
+//                // Set doubleSided so that the user can always see the model
+//                material.isDoubleSided = true
+//
+//                // For some reasons, the Tr value in the OBJ model is not read
+//                // correctly using SCNSceneSource or Model IO. If the OBJ material has
+//                // the d value.
+//                if let transparent = material.transparent.contents as? NSNumber {
+//                    material.transparency = CGFloat(transparent.floatValue)
+//                }
+//            }
+//
+//        }
+//
+//        for childNode in sceneNode.childNodes {
+//            adjustMaterialProperties(sceneNode: childNode)
+//        }
+//    }
     
-    func logSceneSource(_ sceneSource: SCNSceneSource) {
-        if let assetContributor = sceneSource.property(forKey: SCNSceneSourceAssetContributorsKey) as? String {
-            print("Scene Source: Asset Contributor: \(assetContributor)")
-        }
-        
-        if let assetCreatedDate = sceneSource.property(forKey: SCNSceneSourceAssetCreatedDateKey) as? String {
-            print("Scene Source: Asset Created Date: \(assetCreatedDate)")
-        }
-        
-        if let assetModifiedDate = sceneSource.property(forKey: SCNSceneSourceAssetModifiedDateKey) as? String {
-            print("Scene Source: Asset Modified Date: \(assetModifiedDate)")
-        }
-        
-        if let assetUpAxis = sceneSource.property(forKey: SCNSceneSourceAssetUpAxisKey) as? String {
-            print("Scene Source: Asset Up Axis: \(assetUpAxis)")
-        }
-        
-        if let assetUnit = sceneSource.property(forKey: SCNSceneSourceAssetUnitKey) as? String {
-            print("Scene Source: Asset Unit: \(assetUnit)")
-        }
-    }
+//    func logSceneSource(_ sceneSource: SCNSceneSource) {
+//        if let assetContributor = sceneSource.property(forKey: SCNSceneSourceAssetContributorsKey) as? String {
+//            print("Scene Source: Asset Contributor: \(assetContributor)")
+//        }
+//
+//        if let assetCreatedDate = sceneSource.property(forKey: SCNSceneSourceAssetCreatedDateKey) as? String {
+//            print("Scene Source: Asset Created Date: \(assetCreatedDate)")
+//        }
+//
+//        if let assetModifiedDate = sceneSource.property(forKey: SCNSceneSourceAssetModifiedDateKey) as? String {
+//            print("Scene Source: Asset Modified Date: \(assetModifiedDate)")
+//        }
+//
+//        if let assetUpAxis = sceneSource.property(forKey: SCNSceneSourceAssetUpAxisKey) as? String {
+//            print("Scene Source: Asset Up Axis: \(assetUpAxis)")
+//        }
+//
+//        if let assetUnit = sceneSource.property(forKey: SCNSceneSourceAssetUnitKey) as? String {
+//            print("Scene Source: Asset Unit: \(assetUnit)")
+//        }
+//    }
     
-    func logSceneNode(_ sceneNode: SCNNode, level: Int) {
-
-        let currentLevel = max(0 as Int, level)
-        let indentation = String(repeating: "    ", count: currentLevel)
-        let childIndentation = String(repeating: "    ", count: currentLevel + 1)
-        
-        if let name = sceneNode.name {
-            print("\(indentation)SCNNode name: \(name)")
-        } else {
-            print("\(indentation)SCNNode name: <none>")
-        }
-        
-        let (minCoord, maxCoord) = sceneNode.boundingBox
-        print("\(childIndentation)boundingBox = '\(minCoord)', '\(maxCoord)'")
-        
-        if let geometry = sceneNode.geometry {
-            print("\(childIndentation)geometry: \(geometry)")
-            logGeometry(geometry, level: currentLevel + 2)
-        } else {
-            print("\(childIndentation)geometry: <none>")
-        }
-        
-        for childNode in sceneNode.childNodes {
-            logSceneNode(childNode, level: currentLevel + 1)
-        }
-    }
-    
-    func logGeometry(_ geometry: SCNGeometry, level: Int) {
-        let currentLevel = max(0 as Int, level)
-        let indentation = String(repeating: "    ", count: currentLevel)
-        
-        if let name = geometry.name {
-            print("\(indentation)SCNGeometry name: \(name)")
-        } else {
-            print("\(indentation)SCNGeometry name: <none>")
-        }
-        
-        print("\(indentation)Geometry Elements: \(geometry.elements.count)")
-        for (index, geometryElement) in geometry.elements.enumerated() {
-            logGeometryElement(geometryElement, level: currentLevel + 1, prefix: "\(index)")
-        }
-        
-        print("\(indentation)Materials: \(geometry.materials.count)")
-        for (index, material) in geometry.materials.enumerated() {
-            logMaterial(material, level: currentLevel + 1, prefix: "\(index)")
-        }
-    }
-    
-    func logGeometryElement(_ element: SCNGeometryElement, level: Int, prefix: String) {
-        let currentLevel = max(0 as Int, level)
-        let indentation = String(repeating: "    ", count: currentLevel)
-
-        print("\(indentation)\(prefix): \(element)")
-    }
-    
-    func logMaterial(_ material: SCNMaterial, level: Int, prefix: String) {
-        let currentLevel = max(0 as Int, level)
-        let indentation = String(repeating: "    ", count: currentLevel)
-        let childIndentation = String(repeating: "    ", count: currentLevel + 1)
-        
-        if let materialName = material.name {
-            print("\(indentation)\(prefix): name: \(materialName)")
-        } else {
-            print("\(indentation)\(prefix): name: <none>")
-        }
-        
-        print("\(childIndentation)lightingModel: \(material.lightingModel)")
-        print("\(childIndentation)shininess: \(material.shininess)")
-        print("\(childIndentation)fresnelExponent: \(material.fresnelExponent)")
-        print("\(childIndentation)isLitPerPixel: \(material.isLitPerPixel)")
-        print("\(childIndentation)isDoubleSided: \(material.isDoubleSided)")
-        print("\(childIndentation)cullMode: \(material.cullMode)")
-        print("\(childIndentation)blendMode: \(material.blendMode)")
-        print("\(childIndentation)locksAmbientWithDiffuse: \(material.locksAmbientWithDiffuse)")
-        print("\(childIndentation)writesToDepthBuffer: \(material.writesToDepthBuffer)")
-        print("\(childIndentation)readsFromDepthBuffer: \(material.readsFromDepthBuffer)")
-        print("\(childIndentation)colorBufferWriteMask: \(material.colorBufferWriteMask)")
-        print("\(childIndentation)fillMode: \(material.fillMode)")
-        print("\(childIndentation)transparency: \(material.transparency)")
-        logMaterialProperty(material.transparent, level: currentLevel + 1, prefix: "transparent")
-        logMaterialProperty(material.diffuse, level: currentLevel + 1, prefix: "diffuse")
-        logMaterialProperty(material.specular, level: currentLevel + 1, prefix: "specular")
-        logMaterialProperty(material.ambient, level: currentLevel + 1, prefix: "ambient")
-        logMaterialProperty(material.ambientOcclusion, level: currentLevel + 1, prefix: "ambientOcclusion")
-        logMaterialProperty(material.selfIllumination, level: currentLevel + 1, prefix: "selfIllumination")
-        logMaterialProperty(material.metalness, level: currentLevel + 1, prefix: "metalness")
-        logMaterialProperty(material.roughness, level: currentLevel + 1, prefix: "roughness")
-        logMaterialProperty(material.displacement, level: currentLevel + 1, prefix: "displacement")
-        logMaterialProperty(material.normal, level: currentLevel + 1, prefix: "normal")
-        logMaterialProperty(material.reflective, level: currentLevel + 1, prefix: "reflective")
-        logMaterialProperty(material.emission, level: currentLevel + 1, prefix: "emission")
-        
-    }
-    
-    func logMaterialProperty(_ materialProperty: SCNMaterialProperty, level: Int, prefix: String) {
-        let currentLevel = max(0 as Int, level)
-        let indentation = String(repeating: "    ", count: currentLevel)
-        
-        print("\(indentation)\(prefix): \(materialProperty)")
-        
-    }
-    
-    // MARK: FileManagerDelegate
-    
-    func fileManager(_ fileManager: FileManager, shouldRemoveItemAt URL: URL) -> Bool {
-        
-        // For some unknown reasons, the follow commented code caused some textures not
-        // being able to be removed from the previous model and incorrectly apply the
-        // textures to the next model. We reverted back to always return true, but this
-        // causes an error if the user opens a file without anchoring the model on a
-        // plane before closing the file.
-        //if fileManager.fileExists(atPath: URL.absoluteString) {
-        //    return true
-        //} else {
-        //    return false
-        //}
-        return true
-    }
-    
-    func dimension(_ sceneNode: SCNNode) -> SCNVector3 {
-        let (minCoord, maxCoord) = sceneNode.boundingBox
-        return SCNVector3(maxCoord.x - minCoord.x, maxCoord.y - minCoord.y, maxCoord.z - minCoord.z)
-    }
+//    func logSceneNode(_ sceneNode: SCNNode, level: Int) {
+//
+//        let currentLevel = max(0 as Int, level)
+//        let indentation = String(repeating: "    ", count: currentLevel)
+//        let childIndentation = String(repeating: "    ", count: currentLevel + 1)
+//        
+//        if let name = sceneNode.name {
+//            print("\(indentation)SCNNode name: \(name)")
+//        } else {
+//            print("\(indentation)SCNNode name: <none>")
+//        }
+//        
+//        let (minCoord, maxCoord) = sceneNode.boundingBox
+//        print("\(childIndentation)boundingBox = '\(minCoord)', '\(maxCoord)'")
+//        
+//        if let geometry = sceneNode.geometry {
+//            print("\(childIndentation)geometry: \(geometry)")
+//            logGeometry(geometry, level: currentLevel + 2)
+//        } else {
+//            print("\(childIndentation)geometry: <none>")
+//        }
+//        
+//        for childNode in sceneNode.childNodes {
+//            logSceneNode(childNode, level: currentLevel + 1)
+//        }
+//    }
+//    
+//    func logGeometry(_ geometry: SCNGeometry, level: Int) {
+//        let currentLevel = max(0 as Int, level)
+//        let indentation = String(repeating: "    ", count: currentLevel)
+//        
+//        if let name = geometry.name {
+//            print("\(indentation)SCNGeometry name: \(name)")
+//        } else {
+//            print("\(indentation)SCNGeometry name: <none>")
+//        }
+//        
+//        print("\(indentation)Geometry Elements: \(geometry.elements.count)")
+//        for (index, geometryElement) in geometry.elements.enumerated() {
+//            logGeometryElement(geometryElement, level: currentLevel + 1, prefix: "\(index)")
+//        }
+//        
+//        print("\(indentation)Materials: \(geometry.materials.count)")
+//        for (index, material) in geometry.materials.enumerated() {
+//            logMaterial(material, level: currentLevel + 1, prefix: "\(index)")
+//        }
+//    }
+//    
+//    func logGeometryElement(_ element: SCNGeometryElement, level: Int, prefix: String) {
+//        let currentLevel = max(0 as Int, level)
+//        let indentation = String(repeating: "    ", count: currentLevel)
+//
+//        print("\(indentation)\(prefix): \(element)")
+//    }
+//    
+//    func logMaterial(_ material: SCNMaterial, level: Int, prefix: String) {
+//        let currentLevel = max(0 as Int, level)
+//        let indentation = String(repeating: "    ", count: currentLevel)
+//        let childIndentation = String(repeating: "    ", count: currentLevel + 1)
+//        
+//        if let materialName = material.name {
+//            print("\(indentation)\(prefix): name: \(materialName)")
+//        } else {
+//            print("\(indentation)\(prefix): name: <none>")
+//        }
+//        
+//        print("\(childIndentation)lightingModel: \(material.lightingModel)")
+//        print("\(childIndentation)shininess: \(material.shininess)")
+//        print("\(childIndentation)fresnelExponent: \(material.fresnelExponent)")
+//        print("\(childIndentation)isLitPerPixel: \(material.isLitPerPixel)")
+//        print("\(childIndentation)isDoubleSided: \(material.isDoubleSided)")
+//        print("\(childIndentation)cullMode: \(material.cullMode)")
+//        print("\(childIndentation)blendMode: \(material.blendMode)")
+//        print("\(childIndentation)locksAmbientWithDiffuse: \(material.locksAmbientWithDiffuse)")
+//        print("\(childIndentation)writesToDepthBuffer: \(material.writesToDepthBuffer)")
+//        print("\(childIndentation)readsFromDepthBuffer: \(material.readsFromDepthBuffer)")
+//        print("\(childIndentation)colorBufferWriteMask: \(material.colorBufferWriteMask)")
+//        print("\(childIndentation)fillMode: \(material.fillMode)")
+//        print("\(childIndentation)transparency: \(material.transparency)")
+//        logMaterialProperty(material.transparent, level: currentLevel + 1, prefix: "transparent")
+//        logMaterialProperty(material.diffuse, level: currentLevel + 1, prefix: "diffuse")
+//        logMaterialProperty(material.specular, level: currentLevel + 1, prefix: "specular")
+//        logMaterialProperty(material.ambient, level: currentLevel + 1, prefix: "ambient")
+//        logMaterialProperty(material.ambientOcclusion, level: currentLevel + 1, prefix: "ambientOcclusion")
+//        logMaterialProperty(material.selfIllumination, level: currentLevel + 1, prefix: "selfIllumination")
+//        logMaterialProperty(material.metalness, level: currentLevel + 1, prefix: "metalness")
+//        logMaterialProperty(material.roughness, level: currentLevel + 1, prefix: "roughness")
+//        logMaterialProperty(material.displacement, level: currentLevel + 1, prefix: "displacement")
+//        logMaterialProperty(material.normal, level: currentLevel + 1, prefix: "normal")
+//        logMaterialProperty(material.reflective, level: currentLevel + 1, prefix: "reflective")
+//        logMaterialProperty(material.emission, level: currentLevel + 1, prefix: "emission")
+//        
+//    }
+//    
+//    func logMaterialProperty(_ materialProperty: SCNMaterialProperty, level: Int, prefix: String) {
+//        let currentLevel = max(0 as Int, level)
+//        let indentation = String(repeating: "    ", count: currentLevel)
+//        
+//        print("\(indentation)\(prefix): \(materialProperty)")
+//        
+//    }
+//    
+//    // MARK: FileManagerDelegate
+//    
+//    func fileManager(_ fileManager: FileManager, shouldRemoveItemAt URL: URL) -> Bool {
+//        
+//        // For some unknown reasons, the follow commented code caused some textures not
+//        // being able to be removed from the previous model and incorrectly apply the
+//        // textures to the next model. We reverted back to always return true, but this
+//        // causes an error if the user opens a file without anchoring the model on a
+//        // plane before closing the file.
+//        //if fileManager.fileExists(atPath: URL.absoluteString) {
+//        //    return true
+//        //} else {
+//        //    return false
+//        //}
+//        return true
+//    }
+//    
+//    func dimension(_ sceneNode: SCNNode) -> SCNVector3 {
+//        let (minCoord, maxCoord) = sceneNode.boundingBox
+//        return SCNVector3(maxCoord.x - minCoord.x, maxCoord.y - minCoord.y, maxCoord.z - minCoord.z)
+//    }
 }
